# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * واجهة نظام مخزون الأدوية الذكي
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StockStatus = typeof StockStatus[keyof typeof StockStatus];


export const StockStatus = {
  available: 'available',
  low: 'low',
  critical: 'critical',
} as const;

export type MedicinePriority = typeof MedicinePriority[keyof typeof MedicinePriority];


export const MedicinePriority = {
  vital: 'vital',
  essential: 'essential',
  desirable: 'desirable',
} as const;

export type SupplyStatus = typeof SupplyStatus[keyof typeof SupplyStatus];


export const SupplyStatus = {
  draft: 'draft',
  ordered: 'ordered',
  shipped: 'shipped',
  received: 'received',
  delayed: 'delayed',
} as const;

export type AlertSeverity = typeof AlertSeverity[keyof typeof AlertSeverity];


export const AlertSeverity = {
  info: 'info',
  warning: 'warning',
  critical: 'critical',
} as const;

export type DecisionStatus = typeof DecisionStatus[keyof typeof DecisionStatus];


export const DecisionStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Medicine {
  id: number;
  name: string;
  scientificName: string;
  sku: string;
  category: string;
  stock: number;
  unit: string;
  reorderPoint: number;
  safeStock: number;
  dailyConsumption: number;
  status: StockStatus;
  priority: MedicinePriority;
  expiryDate: string;
  location: string;
  predictedRunoutDate: string;
  coverageDays: number;
}

export interface MedicineUpdate {
  /** @minimum 0 */
  stock?: number;
  /** @minimum 0 */
  reorderPoint?: number;
  /** @minimum 0 */
  safeStock?: number;
  priority?: MedicinePriority;
}

export interface Alert {
  id: number;
  medicineId: number;
  medicineName: string;
  title: string;
  message: string;
  severity: AlertSeverity;
  createdAt: string;
  acknowledged: boolean;
}

export interface SupplyOrder {
  id: number;
  orderNumber: string;
  supplier: string;
  medicineId: number;
  medicineName: string;
  quantity: number;
  status: SupplyStatus;
  orderedAt: string;
  expectedArrival: string;
  priority: MedicinePriority;
}

export interface SupplyOrderInput {
  /** @minLength 2 */
  supplier: string;
  medicineId: number;
  medicineName: string;
  /** @minimum 1 */
  quantity: number;
  expectedArrival: string;
  priority: MedicinePriority;
}

export interface SupplyOrderUpdate {
  status: SupplyStatus;
  expectedArrival?: string;
}

export interface Alternative {
  id: number;
  medicineId: number;
  medicineName: string;
  alternativeName: string;
  activeIngredient: string;
  availableStock: number;
  similarity: number;
  decision: DecisionStatus;
  pharmacistNote: string;
}

export type AlternativeDecisionDecision = typeof AlternativeDecisionDecision[keyof typeof AlternativeDecisionDecision];


export const AlternativeDecisionDecision = {
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface AlternativeDecision {
  decision: AlternativeDecisionDecision;
  pharmacistNote?: string;
}

export interface ConsumptionPoint {
  day: string;
  dispensed: number;
  received: number;
}

export interface Dashboard {
  totalMedicines: number;
  availableCount: number;
  lowCount: number;
  criticalCount: number;
  activeAlerts: number;
  openOrders: number;
  stockHealth: number;
  monthlyStockouts: number;
  earlyDetectionRate: number;
  averageDispenseMinutes: number;
  consumptionTrend: ConsumptionPoint[];
  predictedRunouts: Medicine[];
  recentAlerts: Alert[];
  upcomingOrders: SupplyOrder[];
}

export interface WeatherRisk {
  city: string;
  temperature: number;
  condition: string;
  severity: AlertSeverity;
  riskWindow: string;
  recommendation: string;
  affectedOrders: number;
  updatedAt: string;
}

export type ListMedicinesParams = {
search?: string;
status?: StockStatus;
priority?: MedicinePriority;
};