# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * واجهة نظام مخزون الأدوية الذكي
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSeverity } from './alertSeverity';

export interface Alert {
  id: number;
  medicineId: number;
  medicineName: string;
  title: string;
  message: string;
  severity: AlertSeverity;
  createdAt: Date;
  acknowledged: boolean;
}