# lib/api-zod/src/generated/types/alternative.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * واجهة نظام مخزون الأدوية الذكي
 * OpenAPI spec version: 0.1.0
 */
import type { DecisionStatus } from './decisionStatus';

export interface Alternative {
  id: number;
  medicineId: number;
  medicineName: string;
  alternativeName: string;
  activeIngredient: string;
  availableStock: number;
  similarity: number;
  decision: DecisionStatus;
  pharmacistNote: string;
}