# lib/api-zod/src/generated/types/consumptionPoint.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * واجهة نظام مخزون الأدوية الذكي
 * OpenAPI spec version: 0.1.0
 */

export interface ConsumptionPoint {
  day: string;
  dispensed: number;
  received: number;
}