# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * واجهة نظام مخزون الأدوية الذكي
 * OpenAPI spec version: 0.1.0
 */
import type { Alert } from './alert';
import type { ConsumptionPoint } from './consumptionPoint';
import type { Medicine } from './medicine';
import type { SupplyOrder } from './supplyOrder';

export interface Dashboard {
  totalMedicines: number;
  availableCount: number;
  lowCount: number;
  criticalCount: number;
  activeAlerts: number;
  openOrders: number;
  stockHealth: number;
  monthlyStockouts: number;
  earlyDetectionRate: number;
  averageDispenseMinutes: number;
  consumptionTrend: ConsumptionPoint[];
  predictedRunouts: Medicine[];
  recentAlerts: Alert[];
  upcomingOrders: SupplyOrder[];
}