# lib/api-zod/src/generated/types/decisionStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * واجهة نظام مخزون الأدوية الذكي
 * OpenAPI spec version: 0.1.0
 */

export type DecisionStatus = typeof DecisionStatus[keyof typeof DecisionStatus];


export const DecisionStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;