# lib/api-zod/src/generated/types/medicine.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * واجهة نظام مخزون الأدوية الذكي
 * OpenAPI spec version: 0.1.0
 */
import type { MedicinePriority } from './medicinePriority';
import type { StockStatus } from './stockStatus';

export interface Medicine {
  id: number;
  name: string;
  scientificName: string;
  sku: string;
  category: string;
  stock: number;
  unit: string;
  reorderPoint: number;
  safeStock: number;
  dailyConsumption: number;
  status: StockStatus;
  priority: MedicinePriority;
  expiryDate: Date;
  location: string;
  predictedRunoutDate: Date;
  coverageDays: number;
}