# lib/api-zod/src/generated/types/medicinePriority.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * واجهة نظام مخزون الأدوية الذكي
 * OpenAPI spec version: 0.1.0
 */

export type MedicinePriority = typeof MedicinePriority[keyof typeof MedicinePriority];


export const MedicinePriority = {
  vital: 'vital',
  essential: 'essential',
  desirable: 'desirable',
} as const;