# lib/api-zod/src/generated/types/medicineUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * واجهة نظام مخزون الأدوية الذكي
 * OpenAPI spec version: 0.1.0
 */
import type { MedicinePriority } from './medicinePriority';

export interface MedicineUpdate {
  /** @minimum 0 */
  stock?: number;
  /** @minimum 0 */
  reorderPoint?: number;
  /** @minimum 0 */
  safeStock?: number;
  priority?: MedicinePriority;
}