# lib/api-zod/src/generated/types/stockStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * واجهة نظام مخزون الأدوية الذكي
 * OpenAPI spec version: 0.1.0
 */

export type StockStatus = typeof StockStatus[keyof typeof StockStatus];


export const StockStatus = {
  available: 'available',
  low: 'low',
  critical: 'critical',
} as const;