# lib/api-zod/src/generated/types/supplyOrder.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * واجهة نظام مخزون الأدوية الذكي
 * OpenAPI spec version: 0.1.0
 */
import type { MedicinePriority } from './medicinePriority';
import type { SupplyStatus } from './supplyStatus';

export interface SupplyOrder {
  id: number;
  orderNumber: string;
  supplier: string;
  medicineId: number;
  medicineName: string;
  quantity: number;
  status: SupplyStatus;
  orderedAt: Date;
  expectedArrival: Date;
  priority: MedicinePriority;
}