# lib/api-zod/src/generated/types/supplyOrderInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * واجهة نظام مخزون الأدوية الذكي
 * OpenAPI spec version: 0.1.0
 */
import type { MedicinePriority } from './medicinePriority';

export interface SupplyOrderInput {
  /** @minLength 2 */
  supplier: string;
  medicineId: number;
  medicineName: string;
  /** @minimum 1 */
  quantity: number;
  expectedArrival: Date;
  priority: MedicinePriority;
}