# lib/api-zod/src/generated/types/supplyOrderUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * واجهة نظام مخزون الأدوية الذكي
 * OpenAPI spec version: 0.1.0
 */
import type { SupplyStatus } from './supplyStatus';

export interface SupplyOrderUpdate {
  status: SupplyStatus;
  expectedArrival?: Date;
}