# lib/api-zod/src/generated/types/supplyStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * واجهة نظام مخزون الأدوية الذكي
 * OpenAPI spec version: 0.1.0
 */

export type SupplyStatus = typeof SupplyStatus[keyof typeof SupplyStatus];


export const SupplyStatus = {
  draft: 'draft',
  ordered: 'ordered',
  shipped: 'shipped',
  received: 'received',
  delayed: 'delayed',
} as const;