# lib/api-zod/src/generated/types/weatherRisk.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * واجهة نظام مخزون الأدوية الذكي
 * OpenAPI spec version: 0.1.0
 */
import type { AlertSeverity } from './alertSeverity';

export interface WeatherRisk {
  city: string;
  temperature: number;
  condition: string;
  severity: AlertSeverity;
  riskWindow: string;
  recommendation: string;
  affectedOrders: number;
  updatedAt: Date;
}